requirements.txt
scipy==1.3.2
matplotlib==3.1.1
numpy==1.18.1
numpy-stl==2.10.1